backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
pydantic==2.5.3
email-validator>=2.0.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.26.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
tensorflow==2.15.0
prophet==1.1.5
google-ads==23.1.0
facebook-business==19.0.0
ShopifyAPI
ShopifyAPI==12.3.0
google-analytics-data==0.18.4
stripe==7.11.0
sentry-sdk==1.39.2
prometheus-client==0.19.0
loguru==0.7.2
python-dotenv==1.0.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
faker==22.0.0
